API/requirements.txt
Flask
gunicorn==19.6.0
newspaper3k
beautifulsoup4
pickle
flask_cors